GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - master (refs/gitbutler/master)
   - LSProjectTool.xcodeproj/project.pbxproj
   - LSProjectTool/Util/瀑布流/SUPElementsFlowLayout.h
   - Podfile
   - LSProjectTool/Util/获取用户位置/GaoDeMap/LSMKMapViewZoomLevel.m
   - LSProjectTool/lalalal/Nav/LSNavigationController.m
   - LSProjectTool/Util/瀑布流/SUPElementsFlowLayout.m
   - LSProjectTool/Util/瀑布流/SUPVerticalFlowLayout.h
   - LSProjectTool/Util/获取用户位置/GaoDeMap/GaoDeMapLocationService.h
   - LSProjectTool/Util/Tools/ClassExtension/ClassExtension.h
   - LSProjectTool/Util/瀑布流/SUPHorizontalFlowLayout.h
   - LSProjectTool/Util/获取用户位置/GaoDeMap/GaoDeMapLocationService.m
   - LSProjectTool/File/导航栏Nav.h
   - Podfile.lock
   - LSProjectTool/lalalal/Nav/UINavigationBar+iOS11Spacing.h
   - LSProjectTool/Util/瀑布流/SUPVerticalFlowLayout.m
   - LSProjectTool/Util/获取用户位置/GaoDeMap/LSMKMapViewZoomLevel.h
   - LSProjectTool/lalalal/Nav/UINavigationBar+iOS11Spacing.m
   - LSProjectTool/Util/瀑布流/SUPHorizontalFlowLayout.m
   - LSProjectTool/lalalal/BaseViewController/LSRootCollectionView.m

Your previous branch was: refs/heads/master

The sha for that commit was: de7f0e9ae472807a31267de6746600775807bbcc

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch